BF(workaround): install p7zip-full so we do not get stuck testing .gz files
DataLad core has a number of improvements in handling of .gz files in
https://github.com/datalad/datalad/pull/3176
but overall proper solution requires RFing of how all the .gz should be handled
including datalad-archives special remote, so they do not rely on any filename
in the header.  That is for now postponed.
Installing p7zip-full should mitigate the issue for now