RF: annex: Add compatibility kludge to support new create
rev-create, soon to be the main create, doesn't allow configuring the
backend via an argument.  To be compatible with the new create, we
could set datalad.repo.backend to whatever value is configured via a
crawler-specific mechanism.  But it's not clear that any
crawler-specific mechanisms are needed to configure the backend, so
instead warn that the crawler setting isn't honored and point to
datalad.repo.backend.

Re: https://github.com/datalad/datalad/pull/3383#issuecomment-489440554