RF+OPT: make experiment_labels into a lazy eval property and not require full datasets recursion
In case of drop_empty=True get_projects would also assess each project
for having files. Those experiment_labels were assigned as a side effect
within that heavy get_projects function  which was called in per-dataset
crawler.  That lead to unnecessary crawling of the entire pool of projects.

Now with a dedicated call (also called in get_projects) to get a full list
of projects and assign their map we can avoid the waste