MNT: setup: Explicitly specify test dependencies
We assume these are present because they're dependencies of datalad.
However the master branch of datalad no longer supports Python 2, so
it's free to drop its dependency on the external mock library in favor
of unittest.mock, which comes with the standard library as of Python
3.3.

The floor for nose matches the one in datalad's setup.py.

Re: https://github.com/datalad/datalad/pull/3860