ENH(TST): test for testing drop_immediately
Tried to reuse existing test_drop.  It became a pain due to variety of issues,
and still does not work with a VCR tape for some reason (may be underlying process
restarts) in case of drop_immediately.  But otherwise it passes the test, so
I would consider it sufficient for now