TST: Make openfmri pipeline tests run again (kinda)
This comes at the expense of disabling various checks that count the
number of relationship of commits in the various branches.

I looked at the outcome manually, and it looks as if things are OK.

However, I could not figure out what the tests actually require, or how
that was lost. It needs a person with more insight to investigate this.

In any case, I consider it more important to have this code be executed
in v6+ repos (which is pretty much any recent environment), than
maintaining the test behavior.

Please improve!