Add dependency on datalad_deprecated
This extension is co-installable with -core, without breaking it --
regardless of whether or not -core still provides some functionality.

Due to the age of the crawler, it is likely to be affected most by
potential deprecations. Adding this dependency will make it robust
against coming changes, such as the `ls` deprecation
(https://github.com/datalad/datalad/pull/5569).