RF: Adjust for add-archive-content refactoring
github.com/datalad/datalad/pull/6105 refactors add-archive-content
to be a dataset method. This requires changes in datalad-crawlers
use/adaptation of the function. For one, we need to pass a dataset
instance. Secondly, I had to disable the intergrity check of 'annex',
which used to be returned by add-archive-content, but isn't
anymore.